requirements.txt
psycopg2-binary
pandas
requests
python-dotenv
flask 